Extracted template and logic from task component
