+++ src/components/task/create-hashtags.js
@@ -1,0 +1,11 @@
+/**
+ * Creates hashtags list HTML template
+ * @param {Set} tags - set of tags
+ * @return {String} HTML template
+ */
+const createHashTags = (tags) => [...tags]
+  .map((tag) => `<span class="card__hashtag-inner"><span class="card__hashtag-name">#${tag}</span></span>`)
+  .join(``);
+
+
+export {createHashTags};
