+++ src/components/task/index.js
@@ -1,0 +1,8 @@
+import Component from "../component";
+import {template} from "./template";
+
+export default class Task extends Component {
+  constructor(task) {
+    super(template(task));
+  }
+}
